[fix]: replace all uint to uint256

+++ contracts/AmericanBinaryOptions.sol
@@ -15,16 +15,16 @@ import './interfaces/IAmericanBinaryOptions.sol';
  * of precision.
  */
 contract AmericanBinaryOptions is IAmericanBinaryOptions {
-    using SafeMath for uint;
-    using SafeDecimalMath for uint;
+    using SafeMath for uint256;
+    using SafeDecimalMath for uint256;
     using SignedSafeMath for int;
     using SignedSafeDecimalMath for int;
 
-    uint private constant SECONDS_PER_YEAR = 31536000;
+    uint256 private constant SECONDS_PER_YEAR = 31536000;
     /// @dev Internally this library uses 27 decimals of precision
-    uint private constant PRECISE_UNIT = 1e27;
-    uint private constant LN_2_PRECISE = 693147180559945309417232122;
-    uint private constant SQRT_TWOPI = 2506628274631000502415765285;
+    uint256 private constant PRECISE_UNIT = 1e27;
+    uint256 private constant LN_2_PRECISE = 693147180559945309417232122;
+    uint256 private constant SQRT_TWOPI = 2506628274631000502415765285;
     /// @dev Below this value, return 0
     int private constant MIN_CDF_STD_DIST_INPUT =
         (int(PRECISE_UNIT) * -45) / 10; // -4.5
@@ -33,38 +33,38 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
     /// @dev Below this value, the result is always 0
     int private constant MIN_EXP = -63 * int(PRECISE_UNIT);
     /// @dev Above this value the a lot of precision is lost, and uint256s come close to not being able to handle the size
-    uint private constant MAX_EXP = 100 * PRECISE_UNIT;
+    uint256 private constant MAX_EXP = 100 * PRECISE_UNIT;
     /// @dev Value to use to avoid any division by 0 or values near 0
-    uint private constant MIN_T_ANNUALISED = PRECISE_UNIT / SECONDS_PER_YEAR; // 1 second
-    uint private constant MIN_VOLATILITY = PRECISE_UNIT / 10000; // 0.001%
-    uint private constant VEGA_STANDARDISATION_MIN_DAYS = 7 days;
+    uint256 private constant MIN_T_ANNUALISED = PRECISE_UNIT / SECONDS_PER_YEAR; // 1 second
+    uint256 private constant MIN_VOLATILITY = PRECISE_UNIT / 10000; // 0.001%
+    uint256 private constant VEGA_STANDARDISATION_MIN_DAYS = 7 days;
 
     /*
      * Math Operations
      */
 
     /**
-     * @dev Returns absolute value of an int as a uint.
+     * @dev Returns absolute value of an int as a uint256.
      */
-    function abs(int x) public pure override returns (uint) {
-        return uint(x < 0 ? -x : x);
+    function abs(int x) public pure override returns (uint256) {
+        return uint256(x < 0 ? -x : x);
     }
 
     /**
      * @dev Returns the floor of a PRECISE_UNIT (x - (x % 1e27))
      */
-    function floor(uint x) internal pure returns (uint) {
+    function floor(uint256 x) internal pure returns (uint256) {
         return x - (x % PRECISE_UNIT);
     }
 
     /**
      * @dev Returns the natural log of the value using Halley's method.
      */
-    function ln(uint x) internal pure returns (int) {
+    function ln(uint256 x) internal pure returns (int) {
         int res;
         int next;
 
-        for (uint i = 0; i < 8; i++) {
+        for (uint256 i = 0; i < 8; i++) {
             int e = int(exp(res));
             next = res.add(
                 (int(x).sub(e).mul(2)).divideDecimalRoundPrecise(int(x).add(e))
@@ -78,27 +78,27 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
         return res;
     }
 
-    function lnexternal(uint x) external pure returns (int) {
+    function lnexternal(uint256 x) external pure returns (int) {
         return ln(x);
     }
 
     /**
      * @dev Returns the exponent of the value using taylor expansion with range reduction.
      */
-    function exp(uint x) public pure override returns (uint) {
+    function exp(uint256 x) public pure override returns (uint256) {
         if (x == 0) {
             return PRECISE_UNIT;
         }
         require(x <= MAX_EXP, 'cannot handle exponents greater than 100');
 
-        uint k = floor(x.divideDecimalRoundPrecise(LN_2_PRECISE)) /
+        uint256 k = floor(x.divideDecimalRoundPrecise(LN_2_PRECISE)) /
             PRECISE_UNIT;
-        uint p = 2 ** k;
-        uint r = x.sub(k.mul(LN_2_PRECISE));
+        uint256 p = 2 ** k;
+        uint256 r = x.sub(k.mul(LN_2_PRECISE));
 
-        uint _T = PRECISE_UNIT;
+        uint256 _T = PRECISE_UNIT;
 
-        uint lastT;
+        uint256 lastT;
         for (uint8 i = 16; i > 0; i--) {
             _T = _T.multiplyDecimalRoundPrecise(r / i).add(PRECISE_UNIT);
             if (_T == lastT) {
@@ -114,14 +114,14 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
      * @dev Returns the exponent of the value using taylor expansion with range reduction, with support for negative
      * numbers.
      */
-    function exp(int x) public pure override returns (uint) {
+    function exp(int x) public pure override returns (uint256) {
         if (0 <= x) {
-            return exp(uint(x));
+            return exp(uint256(x));
         } else if (x < MIN_EXP) {
             // exp(-63) < 1e-27, so we just return 0
             return 0;
         } else {
-            return PRECISE_UNIT.divideDecimalRoundPrecise(exp(uint(-x)));
+            return PRECISE_UNIT.divideDecimalRoundPrecise(exp(uint256(-x)));
         }
     }
 
@@ -129,8 +129,8 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
      * @dev Returns the square root of the value using Newton's method. This ignores the unit, so numbers should be
      * multiplied by their unit before being passed in.
      */
-    function sqrt(uint x) public pure override returns (uint y) {
-        uint z = (x.add(1)) / 2;
+    function sqrt(uint256 x) public pure override returns (uint256 y) {
+        uint256 z = (x.add(1)) / 2;
         y = x;
         while (z < y) {
             y = z;
@@ -141,7 +141,7 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
     /**
      * @dev Returns the square root of the value using Newton's method.
      */
-    function sqrtPrecise(uint x) internal pure returns (uint) {
+    function sqrtPrecise(uint256 x) internal pure returns (uint256) {
         // Add in an extra unit factor for the square root to gobble;
         // otherwise, sqrt(x * UNIT) = sqrt(x) * sqrt(UNIT)
         return sqrt(x.mul(PRECISE_UNIT));
@@ -150,7 +150,7 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
     /**
      * @dev The standard normal distribution of the value.
      */
-    function stdNormal(int x) internal pure returns (uint) {
+    function stdNormal(int x) internal pure returns (uint256) {
         return
             exp(-x.multiplyDecimalRoundPrecise(x / 2))
                 .divideDecimalRoundPrecise(SQRT_TWOPI);
@@ -176,16 +176,16 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
      * @param rateDecimal The percentage risk free rate + carry cost
      */
     function americanBinaryOptionPrices(
-        uint timeToExpirySec,
-        uint volatilityDecimal,
-        uint spotDecimal,
-        uint strikeDecimal,
+        uint256 timeToExpirySec,
+        uint256 volatilityDecimal,
+        uint256 spotDecimal,
+        uint256 strikeDecimal,
         int rateDecimal
     ) external view override returns (int call) {
-        uint tAnnualised = annualise(timeToExpirySec);
-        uint spotPrecise = spotDecimal.decimalToPreciseDecimal();
-        uint strikePrecise = strikeDecimal.decimalToPreciseDecimal();
-        uint volatilityPrecise = volatilityDecimal.decimalToPreciseDecimal();
+        uint256 tAnnualised = annualise(timeToExpirySec);
+        uint256 spotPrecise = spotDecimal.decimalToPreciseDecimal();
+        uint256 strikePrecise = strikeDecimal.decimalToPreciseDecimal();
+        uint256 volatilityPrecise = volatilityDecimal.decimalToPreciseDecimal();
         int ratePrecise = rateDecimal.decimalToPreciseDecimal();
 
         // 27
@@ -195,8 +195,8 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
         int xi = ratePrecise.divideDecimalRoundPrecise(int(volatilityPrecise)) -
             int(volatilityPrecise) /
             2;
-        uint b = sqrtPrecise(
-            uint(xi.multiplyDecimalRoundPrecise(xi) + 2 * ratePrecise)
+        uint256 b = sqrtPrecise(
+            uint256(xi.multiplyDecimalRoundPrecise(xi) + 2 * ratePrecise)
         );
         int erf1 = erf(
             (int(b.multiplyDecimalRoundPrecise(tAnnualised)) - a)
@@ -239,7 +239,7 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
      * @dev The standard normal cumulative distribution of the value. Only has to operate precisely between -1 and 1 for
      * the calculation of option prices, but handles up to -4 with good accuracy.
      */
-    function stdNormalCDF(int x) internal pure returns (uint) {
+    function stdNormalCDF(int x) internal pure returns (uint256) {
         // Based on testing, errors are ~0.1% at -4, which is still acceptable; and around 0.3% at -4.5.
         // This function seems to become increasingly inaccurate past -5 ( >%5 inaccuracy)
         // At that range, the values are so low at that we will return 0, as it won't affect any usage of this value.
@@ -253,9 +253,9 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
         }
 
         int t1 = int(1e7 + int((2315419 * abs(x)) / PRECISE_UNIT));
-        uint exponent = uint(x.multiplyDecimalRoundPrecise(x / 2));
+        uint256 exponent = uint256(x.multiplyDecimalRoundPrecise(x / 2));
         int d = int((3989423 * PRECISE_UNIT) / exp(exponent));
-        uint prob = uint(
+        uint256 prob = uint256(
             (d *
                 (3193815 +
                     ((-3565638 +
@@ -273,7 +273,9 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
     /**
      * @dev Converts an integer number of seconds to a fractional number of years.
      */
-    function annualise(uint secs) internal pure returns (uint yearFraction) {
+    function annualise(
+        uint256 secs
+    ) internal pure returns (uint256 yearFraction) {
         return secs.divideDecimalRoundPrecise(SECONDS_PER_YEAR);
     }
 
@@ -290,10 +292,10 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
      * @param rate The percentage risk free rate + carry cost
      */
     function d1d2(
-        uint tAnnualised,
-        uint volatility,
-        uint spot,
-        uint strike,
+        uint256 tAnnualised,
+        uint256 volatility,
+        uint256 spot,
+        uint256 strike,
         int rate
     ) internal pure returns (int d1, int d2) {
         // Set minimum values for tAnnualised and volatility to not break computation in extreme scenarios
@@ -325,18 +327,20 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
      * @param d2 Internal coefficient of Black-Scholes
      */
     function _optionPrices(
-        uint tAnnualised,
-        uint spot,
-        uint strike,
+        uint256 tAnnualised,
+        uint256 spot,
+        uint256 strike,
         int rate,
         int d1,
         int d2
-    ) internal pure returns (uint call, uint put) {
-        uint strikePV = strike.multiplyDecimalRoundPrecise(
+    ) internal pure returns (uint256 call, uint256 put) {
+        uint256 strikePV = strike.multiplyDecimalRoundPrecise(
             exp(-rate.multiplyDecimalRoundPrecise(int(tAnnualised)))
         );
-        uint spotNd1 = spot.multiplyDecimalRoundPrecise(stdNormalCDF(d1));
-        uint strikeNd2 = strikePV.multiplyDecimalRoundPrecise(stdNormalCDF(d2));
+        uint256 spotNd1 = spot.multiplyDecimalRoundPrecise(stdNormalCDF(d1));
+        uint256 strikeNd2 = strikePV.multiplyDecimalRoundPrecise(
+            stdNormalCDF(d2)
+        );
 
         // We clamp to zero if the minuend is less than the subtrahend
         // In some scenarios it may be better to compute put price instead and derive call from it depending on which way
@@ -355,15 +359,15 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
      * @param rateDecimal The percentage risk free rate + carry cost
      */
     function optionPrices(
-        uint timeToExpirySec,
-        uint volatilityDecimal,
-        uint spotDecimal,
-        uint strikeDecimal,
+        uint256 timeToExpirySec,
+        uint256 volatilityDecimal,
+        uint256 spotDecimal,
+        uint256 strikeDecimal,
         int rateDecimal
-    ) external pure override returns (uint call, uint put) {
-        uint tAnnualised = annualise(timeToExpirySec);
-        uint spotPrecise = spotDecimal.decimalToPreciseDecimal();
-        uint strikePrecise = strikeDecimal.decimalToPreciseDecimal();
+    ) external pure override returns (uint256 call, uint256 put) {
+        uint256 tAnnualised = annualise(timeToExpirySec);
+        uint256 spotPrecise = spotDecimal.decimalToPreciseDecimal();
+        uint256 strikePrecise = strikeDecimal.decimalToPreciseDecimal();
         int ratePrecise = rateDecimal.decimalToPreciseDecimal();
         (int d1, int d2) = d1d2(
             tAnnualised,
@@ -406,10 +410,10 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
      * @param spot The current price of the base asset
      */
     function _vega(
-        uint tAnnualised,
-        uint spot,
+        uint256 tAnnualised,
+        uint256 spot,
         int d1
-    ) internal pure returns (uint vega) {
+    ) internal pure returns (uint256 vega) {
         return
             sqrtPrecise(tAnnualised).multiplyDecimalRoundPrecise(
                 stdNormal(d1).multiplyDecimalRoundPrecise(spot)
@@ -424,17 +428,17 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
      */
     function _standardVega(
         int d1,
-        uint spot,
-        uint timeToExpirySec
-    ) internal pure returns (uint) {
-        uint tAnnualised = annualise(timeToExpirySec);
+        uint256 spot,
+        uint256 timeToExpirySec
+    ) internal pure returns (uint256) {
+        uint256 tAnnualised = annualise(timeToExpirySec);
 
         timeToExpirySec = timeToExpirySec < VEGA_STANDARDISATION_MIN_DAYS
             ? VEGA_STANDARDISATION_MIN_DAYS
             : timeToExpirySec;
-        uint daysToExpiry = (timeToExpirySec.mul(PRECISE_UNIT)) / 1 days;
-        uint thirty = 30 * PRECISE_UNIT;
-        uint normalisationFactor = sqrtPrecise(
+        uint256 daysToExpiry = (timeToExpirySec.mul(PRECISE_UNIT)) / 1 days;
+        uint256 thirty = 30 * PRECISE_UNIT;
+        uint256 normalisationFactor = sqrtPrecise(
             thirty.divideDecimalRoundPrecise(daysToExpiry)
         ).div(100);
         return
@@ -452,10 +456,10 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
      * @param rateDecimal The percentage risk free rate + carry cost
      */
     function pricesDeltaStdVega(
-        uint timeToExpirySec,
-        uint volatilityDecimal,
-        uint spotDecimal,
-        uint strikeDecimal,
+        uint256 timeToExpirySec,
+        uint256 volatilityDecimal,
+        uint256 spotDecimal,
+        uint256 strikeDecimal,
         int rateDecimal
     )
         external
@@ -463,8 +467,8 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
         override
         returns (IAmericanBinaryOptions.PricesDeltaStdVega memory)
     {
-        uint tAnnualised = annualise(timeToExpirySec);
-        uint spotPrecise = spotDecimal.decimalToPreciseDecimal();
+        uint256 tAnnualised = annualise(timeToExpirySec);
+        uint256 spotPrecise = spotDecimal.decimalToPreciseDecimal();
 
         (int d1, int d2) = d1d2(
             tAnnualised,
@@ -473,7 +477,7 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
             strikeDecimal.decimalToPreciseDecimal(),
             rateDecimal.decimalToPreciseDecimal()
         );
-        (uint callPrice, uint putPrice) = _optionPrices(
+        (uint256 callPrice, uint256 putPrice) = _optionPrices(
             tAnnualised,
             spotPrecise,
             strikeDecimal.decimalToPreciseDecimal(),
@@ -481,7 +485,7 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
             d1,
             d2
         );
-        uint v = _standardVega(d1, spotPrecise, timeToExpirySec);
+        uint256 v = _standardVega(d1, spotPrecise, timeToExpirySec);
         (int callDelta, int putDelta) = _delta(d1);
 
         return
@@ -495,10 +499,10 @@ contract AmericanBinaryOptions is IAmericanBinaryOptions {
     }
 
     function mockCalculation(
-        uint coverage,
-        uint duration,
-        uint risk
-    ) external pure override returns (uint price) {
+        uint256 coverage,
+        uint256 duration,
+        uint256 risk
+    ) external pure override returns (uint256 price) {
         return 1e18;
     }
 }
